messaging_app/requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.2.1
django-cors-headers 4.7.0
django-environ      0.12.0
djangorestframework 3.16.0
drf-nested-routers  0.94.2
pip                 22.0.2
setuptools          59.6.0
sqlparse            0.5.3
typing_extensions   4.13.2
